use bpkg to install JSON.sh and fix the .bin symlink exists error

+++ source.sh
@@ -44,6 +44,7 @@ add_to_PATH ${DOTS_HOME}/bin
 add_to_PATH .
 mkdir -p ~/bin
 add_to_PATH ~/bin
-ln -s ~/bin ~/.bin
+[[ ! -L ~/.bin ]] && ln -s ~/bin ~/.bin
+add_to_PATH ~/deps/bin # see install/shell
 
 export PATH # concludes what the PATH will be
